# openapi.yaml
openapi: 3.0.3
info:
  title: Vocalytics API
  version: 1.0.0
  description: YouTube comment analytics and reply automation API
servers:
  - url: https://vocalytics-alpha.vercel.app
    description: Production
  - url: http://localhost:3000
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /api/youtube/videos:
    get:
      summary: List user's uploaded videos
      parameters:
        - name: mine
          in: query
          schema:
            type: boolean
            default: true
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: List of videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserVideo'
        '403':
          description: YouTube not connected
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
  /api/analysis/{videoId}:
    post:
      summary: Run sentiment analysis on a video
      parameters:
        - name: videoId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResult'
        '402':
          $ref: '#/components/responses/Paywall'
    get:
      summary: Get latest analysis for a video
      parameters:
        - name: videoId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResult'
        '404':
          description: No analysis found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
...
  /api/analysis/trends:
    get:
      summary: Get sentiment trends over time
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 365
            default: 90
      responses:
        '200':
          description: Trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrendPoint'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT token
  schemas:
...
    UserVideo:
      type: object
      required: [videoId, title, stats]
      properties:
        videoId:
          type: string
        title:
          type: string
        thumbnailUrl:
          type: string
        publishedAt:
          type: string
          format: date-time
        stats:
          type: object
          properties:
            viewCount:
              type: integer
            likeCount:
              type: integer
            commentCount:
              type: integer
    AnalysisResult:
      type: object
      required: [videoId, analyzedAt, sentiment, score]
      properties:
        videoId:
          type: string
        analyzedAt:
          type: string
          format: date-time
        sentiment:
          type: object
          required: [pos, neu, neg]
          properties:
            pos:
              type: number
            neu:
              type: number
            neg:
              type: number
        score:
          type: number
          description: Normalized positivity score (pos - neg)
        topPositive:
          type: array
          items:
            type: object
        topNegative:
          type: array
          items:
            type: object
        summary:
          type: string
...
    TrendPoint:
      type: object
      required: [date, avgScore]
      properties:
        date:
          type: string
          format: date-time
        avgScore:
          type: number
          description: Average sentiment score for this date
  responses:
    Paywall:
      description: Free tier limit exceeded
      content:
        application/json:
          schema:
            type: object
            required: [code, reason, feature, upgradeUrl]
            properties:
              code:
                type: string
                enum: [PAYWALL]
              reason:
                type: string
                enum: [FREE_TIER_EXCEEDED]
              feature:
                type: string
                enum: [analyze, reply]
              upgradeUrl:
                type: string
                format: uri
              manageUrl:
                type: string
                format: uri
              limits:
                type: object
                properties:
                  weeklyAnalyze:
                    type: integer
                  dailyReply:
                    type: integer
              usage:
                type: object
                properties:
                  commentsAnalyzed:
                    type: integer
                  repliesGenerated:
                    type: integer